Moved Request object to a new file, fixed WorldUuid parameter, implemented builder chaining
(cherry picked from commit 864b66901006ec7e6777b00f51c599a2f3cd5eaf)
